encjapanese.c closer

+++ libs/RVL_SDK/include/revolution/enc/encutility.h
@@ -33,7 +33,7 @@ typedef struct ENCContext {
 
 ENCResult ENCiCheckParameters(BOOL destNotNull, u32* destLength, s32* newDestLength, BOOL* destParamsValid,
 BOOL srcNotNull, u32* srcLength, s32* newSrcLength, BOOL* srcParamsValid);
-u32 ENCiCheckBreakType(u8 char1, u8 char2);
+u32 ENCiCheckBreakType(u32 char1, u32 char2);
 u32 ENCiWriteBreakType(u8* dest, u32 size, ENCBreakType breakType, BOOL write);
 
 #ifdef __cplusplus
